Add error checking and validation
Prior to this, the script would fail to run on my Mac because the mktemp
command in OSX requires a template string. Also, there was no error
checking or validation, so the script continued to run even when it
failed from the start.

Also, made the cleanup steps be run with an EXIT trap just to ensure
that temporary files don't get left behind.