Fix trap implementation and verify module download
Prior to this, there was no validation that the module downloaded
successfully before trying to generate the Puppetfile.

Also, the cleanup function was defined at the bottom of the script which
caused the exit trap to not execute the cleanup if exiting mid-script
because the function had not been parsed yet.